Corrected styles and other stuff
