+++ src/app/pages/auth/pages/register/register.component.scss
@@ -2,18 +2,20 @@
 @import 'src/styles/variables/breakpoints';
 
 :host {
+  width: 100%;
+  height: 100%;
   display: flex;
   flex-direction: column;
   justify-content: center;
   align-items: center;
   text-align: center;
+  gap: 20px;
 }
 
 .title {
   color: map-get($map: $ds-grey, $key: 200);
   font-size: 32px;
   text-transform: uppercase;
-  margin-top: 30px;
   font-weight: 400;
 }
 
@@ -27,7 +29,6 @@
   flex-direction: column;
   gap: 20px;
   color: map-get($ds-utils, font-color);
-  margin-top: 50px;
   width: 300px;
 }
 
@@ -43,12 +44,16 @@ mat-label {
 
 .buttons {
   display: flex;
-  flex-direction: row;
-  gap: 50px;
+  flex-direction: column;
+  gap: 10px;
   justify-content: center;
   align-items: baseline;
-  margin-top: 40px;
-  margin-bottom: 30px;
+
+  @media (min-width:$screen-1024) {
+    flex-direction: row;
+    align-items: center;
+    gap: 20px
+  }
 }
 
 .action__button {
@@ -59,17 +64,4 @@ mat-label {
   font-size: 25px;
   width: 215px;
   height: 60px;
-}
-
-@media screen and (max-width:$screen-428) {
-  .title {
-    margin-top: 10px;
-  }
-
-  .buttons {
-    flex-direction: column;
-    gap: 30px;
-    align-items: center;
-    margin-top: 30px;
-  }
 }
