+++ src/app/pages/auth/services/registerFormCreator/register-form-creator.service.ts
@@ -1,20 +1,11 @@
 import { Injectable } from '@angular/core';
-import { AbstractControl, FormBuilder, ValidatorFn, Validators } from '@angular/forms';
+import { FormBuilder, Validators } from '@angular/forms';
+import { PasswordValidator } from '@pages/auth/shared/same-password/same-password.validator';
 
 @Injectable()
 export class RegisterFormCreatorService {
 
   constructor(private fb: FormBuilder) { }
-
-  private passwordValidator: ValidatorFn = (control: AbstractControl): { [key: string]: boolean } | null => {
-    const password = control.get('password');
-    const repeatPassword = control.get('repeatPassword');
-
-    if (!password || !repeatPassword) {
-      return null;
-    }
-    return password.value === repeatPassword.value ? null : { passwordsNotSame: true };
-  };
 
   createForm(){
     return this.fb.group({
@@ -22,6 +13,6 @@ export class RegisterFormCreatorService {
       email: [null, [Validators.email, Validators.required, Validators.maxLength(255)]],
       password: [null, [Validators.required, Validators.minLength(8), Validators.maxLength(255)]],
       repeatPassword: [null, [Validators.required, Validators.minLength(8), Validators.maxLength(255)]]
-    }, {validators: [this.passwordValidator]});
+    }, {validators: [PasswordValidator]});
   }
 }
