+++ src/app/pages/landing-page/landing-page.component.html
@@ -1,6 +1,8 @@
 <h2 class="welcome">Welcome to</h2>
-<ds-logo class="title" fontSize="95px"/>
-<h3 class="description">The perfect place where you can find a dream team mate that you deserve.</h3>
+<div class="wrapper">
+  <ds-logo class="title" />
+  <h3 class="description">The perfect place where you can find a dream team mate that you deserve.</h3>
+</div>
 <div class="actions">
   <a
     class="action__button"
