+++ src/app/pages/landing-page/landing-page.component.scss
@@ -2,44 +2,90 @@
 @import 'src/styles/variables/breakpoints';
 
 :host {
+  gap: 30px;
+  width: 100%;
+  height: 100vh;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  flex-direction: column;
+  text-align: center;
+
+  @media (min-width:$screen-440) {
+    gap: 40px;
+  }
+}
+
+.wrapper {
   display: flex;
   flex-direction: column;
-  gap: 50px;
-  width: 50%;
+  gap: 60px;
+  justify-content: center;
+  align-items: center;
   font-weight: 400;
-  margin: auto;
-  text-align: center;
+
+  @media (min-width:$screen-440) {
+    gap: 80px;
+  }
 }
 
 .welcome {
-  margin-top: 300px;
-  font-size: 55px;
+  font-size: 32px;
   text-transform: uppercase;
   font-weight: 400;
   color: map-get($map: $ds-grey, $key: 200);
+
+  @media (min-width:$screen-440) and (max-width:$screen-840) {
+    font-size: 45px;
+  }
+
+  @media (min-width:$screen-840) {
+    font-size: 55px;
+  }
 }
 
 .title {
-  font-size: 95px;
+  font-size: 60px;
+
+  @media (min-width:$screen-440) and (max-width:$screen-840) {
+    font-size: 75px;
+  }
+
+  @media (min-width:$screen-840) {
+    font-size: 95px;
+  }
 }
 
 .description {
-  line-height: 45px;
-  font-size: 35px;
-  margin: auto;
-  margin-top: 10px;
-  width: 700px;
-  height: 40px;
+  line-height: 35px;
+  font-size: 30px;
+  width: 85%;
   color: map-get($map: $ds-grey, $key: 200);
+
+  @media (min-width:$screen-440) and (max-width:$screen-840){
+    line-height: 40px;
+    font-size: 32px;
+    max-width: 400px;
+  }
+
+  @media (min-width:$screen-840) {
+    line-height: 40px;
+    font-size: 35px;
+    max-width: 700px;
+  }
 }
 
 .actions {
   display: flex;
-  flex-direction: row;
-  gap: 80px;
-  justify-content: center;
-  align-items: baseline;
-  margin-top: 100px;
+  flex-direction: column;
+  gap: 30px;
+  align-items: center;
+
+  @media (min-width:$screen-840) {
+    flex-direction: row;
+    justify-content: center;
+    gap: 80px;
+  }
 }
 
 .action__button {
@@ -50,36 +96,4 @@
   font-size: 25px;
   width: 215px;
   height: 60px;
-}
-
-
-@media screen and (max-width:$screen-428) {
-  :host {
-    gap: 20px;
-    width: 100%;
-  }
-
-  .welcome {
-    margin-top: 75px;
-    font-size: 32px;
-  }
-
-  .title {
-    font-size: 60px;
-  }
-
-  .description {
-    line-height: 40px;
-    font-size: 30px;
-    margin-top: 70px;
-    width: 85%;
-    height: 132px;
-  }
-
-  .actions {
-    flex-direction: column;
-    align-items: center;
-    gap: 40px;
-    margin-top: 60px;
-  }
 }
