+++ src/app/shared/ui/logo/logo.component.scss
@@ -1,4 +1,4 @@
-@import '../../../../styles/variables/palette.scss';
+@import 'src/styles/variables/palette.scss';
 
 :host {
   display: block;
