+++ src/styles/components/card.scss
@@ -1,5 +1,5 @@
 //card
-@import '../variables/palette';
+@import 'src/styles/variables/palette';
 
 .ds__card {
   background-color: map-get($map: $ds-grey, $key: 450);
