+++ src/styles/styles.scss
@@ -50,14 +50,13 @@ $frontend-theme: mat.define-light-theme((color: (primary: $frontend-primary,
 
 html,
 body {
-  min-height: 100%;
+  // min-height: 100%;
   background-color: map-get($map: $ds-utils, $key: background-color);
   color: map-get($map: $ds-utils, $key: font-color);
 }
 
 body {
   margin: 0;
-  // font-family: 'Roboto Slab', serif;
   font-family: 'Space Grotesk', sans-serif;
   font-size: 14px;
 }
