+++ src/styles/variables/breakpoints.scss
@@ -1,6 +1,6 @@
 $screen-360: 360px;
 $screen-400: 400px;
-$screen-428: 428px;
+$screen-440: 440px;
 $screen-550: 550px;
 $screen-600: 600px;
 $screen-720: 720px;
